requirements.txt
python-slackclient==1.0.5
requests==2.13.0
schedule==0.4.1